pyproject.toml
[project]
name = "scoremanager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "supabase>=2.27.2",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["."]
include = ["supabase.scripts*"]

[project.scripts]
register = "supabase.scripts.register_actions:run"
